Move EmailUtil to util package
